refactor: add AppError in backend services

+++ backend/src/services/apontamentos.ts
@@ -1,34 +1,69 @@
 import { prisma } from "../config/prisma";
 import { Apontamento, Prisma } from "@prisma/client";
+import { AppError } from "../errors/appError";
 
 export class ApontamentoService{
 
     async pegarTodosApontamentos(): Promise<Apontamento[]>{
-            return await prisma.apontamento.findMany();
+            const todosApontamentos = await prisma.apontamento.findMany();
+
+            if(todosApontamentos.length === 0){
+                throw new AppError("Liste de apontamentos vazia", 404)
+            }
+            return todosApontamentos
     };
 
     async pegarApontamentoPorId(id: number): Promise<Apontamento | null>{
-            return await prisma.apontamento.findUnique({
+            const idApontamentoPego = await prisma.apontamento.findUnique({
             where: {id}
         });
+
+        if(idApontamentoPego){
+            throw new AppError("Id de apontamento não encontrado", 404)
+        }
+        return idApontamentoPego
     };
 
     async criarApontamento(apontamentoData: Prisma.ApontamentoCreateInput): Promise<Apontamento>{
-            return await prisma.apontamento.create({
+            const apontamentoCriado = await prisma.apontamento.create({
             data: apontamentoData
         });
+
+        if(!apontamentoCriado){
+            throw new AppError("Não foi possível criar o apontamento")
+        }
+        return apontamentoCriado
     };
 
     async atualizarApontamento(id: number, apontamentoData: Prisma.ApontamentoUpdateInput): Promise<Apontamento> {
-            return await prisma.apontamento.update({
+        
+        const apontamentoExiste = await prisma.usuario.findUnique({
+            where: { id }
+        })
+
+        if(!apontamentoExiste){
+            throw new AppError("Id de apontamento não encontrado", 404)
+        }
+        
+        const atualizaApontamento = await prisma.apontamento.update({
                 where: { id },
                 data: apontamentoData
         });
-    ;}
+        return atualizaApontamento
+    };
         
     async deletarApontamento(id: number): Promise<Apontamento> {
-            return await prisma.apontamento.delete({
+            const apontamentoExiste = await prisma.usuario.findUnique({
+                where: { id }
+            })
+            
+            if(!apontamentoExiste){
+                throw new AppError("Id de apontamento não encontrado", 404)
+            }
+
+            const deletaApontamento = await prisma.apontamento.delete({
                 where: { id }
         });
+            return deletaApontamento
     };
 };
